Extract UserResearchRecruitmentController
In order to keep this concern out of the RootController.
