+++ app/controllers/root_controller.rb
@@ -1,6 +1,4 @@
 class RootController < ApplicationController
-  USER_RESEARCH_RECRUITMENT_FORM_URL = "https://docs.google.com/forms/d/1Bdu_GqOrSR4j6mbuzXkFTQg6FRktRMQc8Y-q879Mny8/viewform".freeze
-
   layout "admin_layout"
 
   include UserPermissionsControllerMethods
@@ -15,16 +13,6 @@ class RootController < ApplicationController
 
   def signin_required
     @application = ::Doorkeeper::Application.find_by(id: session.delete(:signin_missing_for_application))
-  end
-
-  def dismiss_user_research_recruitment_banner
-    cookies[:dismiss_user_research_recruitment_banner] = true
-    redirect_to root_path
-  end
-
-  def user_research_recruitment_form
-    current_user.update!(user_research_recruitment_banner_hidden: true)
-    redirect_to USER_RESEARCH_RECRUITMENT_FORM_URL, allow_other_host: true
   end
 
 private
